backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
aiohttp>=3.10.0
anthropic>=0.40.0
openai>=1.51.0
pinecone>=5.0.0
python-dotenv>=1.0.0
